+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/extent/filter/MaskFilter.java
@@ -9,7 +9,7 @@ import com.fastasyncworldedit.core.internal.simd.VectorizedMask;
 import com.fastasyncworldedit.core.queue.Filter;
 import com.sk89q.worldedit.function.mask.AbstractExtentMask;
 import com.sk89q.worldedit.function.mask.Mask;
-import jdk.incubator.vector.ShortVector;
+import jdk.incubator.vector.IntVector;
 import jdk.incubator.vector.VectorMask;
 import jdk.incubator.vector.VectorOperators;
 import jdk.incubator.vector.VectorSpecies;
@@ -84,14 +84,14 @@ public class MaskFilter<T extends Filter> extends DelegateFilter<T> {
         }
 
         @Override
-        public void applyVector(final VectorFacade get, final VectorFacade set, final VectorMask<Short> mask) {
+        public void applyVector(final VectorFacade get, final VectorFacade set, final VectorMask<Integer> mask) {
             final T parent = getParent();
-            final VectorSpecies<Short> species = mask.vectorSpecies();
-            VectorMask<Short> masked = this.vectorizedMask.compareVector(set, get, species);
-            ShortVector before = set.getOrZero(masked.vectorSpecies());
+            final VectorSpecies<Integer> species = mask.vectorSpecies();
+            VectorMask<Integer> masked = this.vectorizedMask.compareVector(set, get, species);
+            IntVector before = set.getOrZero(masked.vectorSpecies());
             parent.applyVector(get, set, mask.and(masked));
-            ShortVector after = set.getOrZero(masked.vectorSpecies());
-            VectorMask<Short> changed = after.compare(VectorOperators.NE, before);
+            IntVector after = set.getOrZero(masked.vectorSpecies());
+            VectorMask<Integer> changed = after.compare(VectorOperators.NE, before);
             this.changes.getAndAdd(changed.trueCount());
         }
