+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/internal/simd/VectorizedCharFilterBlock.java
@@ -3,7 +3,7 @@ package com.fastasyncworldedit.core.internal.simd;
 import com.fastasyncworldedit.core.extent.filter.block.CharFilterBlock;
 import com.fastasyncworldedit.core.queue.Filter;
 import com.sk89q.worldedit.extent.Extent;
-import jdk.incubator.vector.ShortVector;
+import jdk.incubator.vector.IntVector;
 import jdk.incubator.vector.VectorMask;
 import jdk.incubator.vector.VectorSpecies;
 
@@ -23,13 +23,13 @@ public class VectorizedCharFilterBlock extends CharFilterBlock {
         if (!(filter instanceof VectorizedFilter vecFilter)) {
             throw new IllegalStateException("Unexpected VectorizedCharFilterBlock " + filter);
         }
-        final VectorSpecies<Short> species = ShortVector.SPECIES_PREFERRED;
+        final VectorSpecies<Integer> species = IntVector.SPECIES_PREFERRED;
         VectorFacade setFassade = new VectorFacade(this.set);
         setFassade.setLayer(this.layer);
         VectorFacade getFassade = new VectorFacade(this.get);
         getFassade.setLayer(this.layer);
         getFassade.setData(this.getArr);
-        VectorMask<Short> affectAll = species.maskAll(true);
+        VectorMask<Integer> affectAll = species.maskAll(true);
         for (int i = startY << 8; i < ((endY + 1) << 8) - 1; i += species.length()) {
             setFassade.setIndex(i);
             getFassade.setIndex(i);
